+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Changelog
 
 ## 0.9.4
+
+### Added
+
 - `LineSegmentExtensions.Intersections(this IList<Line> lines)`
 - `Elements.Search.DistanceComparer`
 - `Elements.Search.DirectionComparer`
 - `Elements.Search.Network<T>`
+- `ModelExtensions.AllElementsOfType<T>(this Dictionary<string, Model> models, string modelName)`
+
+### Changed
+
+### Fixed
+
 
 ## 0.9.3
 
@@ -25,10 +34,8 @@
 ### Fixed
 
 - Deduplicate catalog names during code generation.
-  <<<<<<< HEAD
-- # Fix some issues with code generation and deserialization of `Vector3` and `Mesh` types.
+- Fix some issues with code generation and deserialization of `Vector3` and `Mesh` types.
 - Fixed an issue where GLTFs would occasionally be generated with incorrect vertex normals.
-  > > > > > > > 3a8c28c8c385007084e1160636a595d7faaaab55
 
 ## 0.9.2